# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "description": "accessToken · refreshToken · user를 반환합니다. 이후 온보딩 설문(`/onboarding`)을 이어가세요.",
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "이메일 회원가입",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "description": "accessToken · refreshToken · user 반환. Swagger Authorize에 accessToken을 넣으면 됩니다.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "이메일 로그인",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "refreshToken으로 새 accessToken을 발급받습니다.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "액세스 토큰 갱신",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "해당 refreshToken을 폐기합니다.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "로그아웃",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/v1/auth/join-by-invite": {
      "post": {
        "description": "계정이 없는 친구가 초대 링크로 들어올 때 사용. 게스트 유저를 만들고 해당 여행방에 바로 입장시킵니다.",
        "operationId": "AuthController_joinByInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinByInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthJoinByInviteDto"
                }
              }
            }
          }
        },
        "summary": "초대코드로 게스트 입장",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/v1/auth/oauth/kakao": {
      "post": {
        "description": "카카오 SDK에서 받은 accessToken을 넘기면 Tourmate 토큰으로 교환합니다.",
        "operationId": "AuthController_kakao",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KakaoOAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "카카오 로그인",
        "tags": [
          "인증"
        ]
      }
    },
...
    "/api/v1/quiz/status": {
      "get": {
        "operationId": "QuizController_getStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테스트 진행·완료 상태",
        "tags": [
          "두리 테스트"
        ]
      }
    },
    "/api/v1/quiz/me": {
      "get": {
        "description": "완료된 테스트의 travelType · axes · preferences를 반환합니다.",
        "operationId": "QuizController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 최신 성향 조회",
        "tags": [
          "두리 테스트"
        ]
      }
    },
    "/api/v1/quiz/sessions": {
      "post": {
        "description": "새 in_progress 세션을 만들고 이전 isLatest를 해제합니다.",
        "operationId": "QuizController_createSession",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테스트 세션 생성",
        "tags": [
          "두리 테스트"
        ]
      }
    },
    "/api/v1/quiz/sessions/{sessionId}": {
      "patch": {
        "description": "보낸 responses 필드를 기존 세션에 merge합니다.",
        "operationId": "QuizController_patchSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchQuizSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "응답 중간 저장",
        "tags": [
          "두리 테스트"
        ]
      }
    },
    "/api/v1/quiz/sessions/{sessionId}/complete": {
      "post": {
        "description": "FE가 계산한 챕터 점수(responses)를 저장하고 TravelType·axes·preferences를 갱신합니다.",
        "operationId": "QuizController_completeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteQuizSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "테스트 완료 → 성향 저장",
        "tags": [
          "두리 테스트"
        ]
      }
    },
    "/api/v1/quiz/submit": {
      "post": {
        "deprecated": true,
        "description": "신규 플로우는 `/quiz/sessions`를 사용하세요.",
        "operationId": "QuizController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSubmitResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[레거시] 한 방 제출",
        "tags": [
          "두리 테스트"
        ]
      }
    },
...
    "/api/v1/rooms/{roomId}/duri/reflect-preferences": {
      "post": {
        "operationId": "DuriController_reflectPreferences",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "멤버 취향 반영 (compatibility 재사용)",
        "tags": [
          "두리 도우미"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/duri/analysis-report": {
      "post": {
        "description": "같은 DAY의 인접 일정만 분석하며, 자동차 경로·시간 겹침·사용자 입력 선호처럼 근거가 있는 결과만 반환합니다.",
        "operationId": "DuriController_createReport",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReportDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "검증 가능한 데이터 기반 일정 리포트 생성",
        "tags": [
          "두리 도우미"
        ]
      }
    },
    "/api/v1/rooms/{roomId}/duri/analysis-report/latest": {
      "get": {
        "operationId": "DuriController_latestReport",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReportDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "최신 분석 리포트 조회",
        "tags": [
          "두리 도우미"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tourmate API",
    "description": "TripMatch / Tourmate backend REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "example": "윤지",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "example": "password1234",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "nickname",
          "email",
          "password"
        ]
      },
...
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665abc123def456789012345"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "nickname": {
            "type": "string",
            "example": "윤지"
          },
          "profileImageUrl": {
            "type": "object",
            "example": "https://example.com/me.jpg",
            "nullable": true
          },
          "travelType": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TravelTypeDto"
              }
            ]
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": false
          },
          "isGuest": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "nickname",
          "onboardingCompleted",
          "isGuest"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "POST /auth/login 또는 /auth/signup 응답의 refreshToken"
          },
          "operationId": {
            "type": "string",
            "example": "8fb9d9a8-71c3-42fc-9690-c18e274f3122",
            "format": "uuid",
            "description": "같은 갱신 요청의 재시도에 유지하는 UUID. 응답 유실 복구와 일회성 토큰 회전을 함께 보장합니다."
          }
        },
        "required": [
          "refreshToken",
          "operationId"
        ]
      },
      "RefreshTokenBodyDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "POST /auth/login 또는 /auth/signup 응답의 refreshToken"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "SuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "JoinByInviteDto": {
        "type": "object",
        "properties": {
          "inviteCode": {
            "type": "string",
            "example": "ABCD1234",
            "description": "여행방 초대 코드 (8자)"
          },
          "nickname": {
            "type": "string",
            "example": "게스트윤지",
            "minLength": 2
          }
        },
        "required": [
          "inviteCode",
          "nickname"
        ]
      },
      "AuthJoinByInviteDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIs..."
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "roomId": {
            "type": "string",
            "example": "665abc123def456789012345"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user",
          "roomId"
        ]
      },
      "KakaoOAuthDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "kakao_access_token_from_sdk",
            "description": "Kakao SDK에서 발급받은 access token"
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "QuizStatusDto": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "boolean",
            "example": true
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": true
          },
          "sessionId": {
            "type": "object",
            "example": "665abc123def456789012345",
            "nullable": true
          },
          "status": {
            "type": "string",
            "example": "in_progress",
            "enum": [
              "in_progress",
              "completed"
            ],
            "nullable": true
          },
          "answeredCount": {
            "type": "number",
            "example": 3
          },
          "totalQuestions": {
            "type": "number",
            "example": 6
          },
          "totalSteps": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "completed",
          "onboardingCompleted",
          "answeredCount",
          "totalQuestions",
          "totalSteps"
        ]
      },
...
      "PartialQuizResponsesDto": {
        "type": "object",
        "properties": {
          "surveyVersion": {
            "type": "number",
            "example": 9
          },
          "algorithmVersion": {
            "type": "number",
            "example": 9
          },
          "challengeStyle": {
            "$ref": "#/components/schemas/ChallengeStyleChapterDto"
          },
          "accommodation": {
            "$ref": "#/components/schemas/AccommodationChapterDto"
          },
          "stamina": {
            "$ref": "#/components/schemas/StaminaChapterDto"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetChapterDto"
          },
          "discovery": {
            "$ref": "#/components/schemas/DiscoveryChapterDto"
          }
        }
      },
      "PatchQuizSessionDto": {
        "type": "object",
        "properties": {
          "responses": {
            "$ref": "#/components/schemas/PartialQuizResponsesDto"
          }
        },
        "required": [
          "responses"
        ]
      },
      "QuizResponsesDto": {
        "type": "object",
        "properties": {
          "surveyVersion": {
            "type": "number",
            "example": 9
          },
          "algorithmVersion": {
            "type": "number",
            "example": 9
          },
          "challengeStyle": {
            "$ref": "#/components/schemas/ChallengeStyleChapterDto"
          },
          "accommodation": {
            "$ref": "#/components/schemas/AccommodationChapterDto"
          },
          "stamina": {
            "$ref": "#/components/schemas/StaminaChapterDto"
          },
          "budget": {
            "$ref": "#/components/schemas/BudgetChapterDto"
          },
          "discovery": {
            "$ref": "#/components/schemas/DiscoveryChapterDto"
          }
        },
        "required": [
          "surveyVersion",
          "algorithmVersion",
          "challengeStyle",
          "accommodation",
          "stamina",
          "budget",
          "discovery"
        ]
      },
      "CompleteQuizSessionDto": {
        "type": "object",
        "properties": {
          "responses": {
            "description": "전체 챕터 응답. FE가 점수까지 계산해서 보냄.",
            "allOf": [
              {
                "$ref": "#/components/schemas/QuizResponsesDto"
              }
            ]
          }
        },
        "required": [
          "responses"
        ]
      },
      "SubmitQuizDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "object",
            "description": "레거시 문항 ID → 선택지 ID. 신규는 /quiz/sessions 사용.",
            "example": {
              "q1": "q1_a"
            }
          }
        },
        "required": [
          "answers"
        ]
      },
      "QuizSubmitResultDto": {
        "type": "object",
        "properties": {
          "travelType": {
            "$ref": "#/components/schemas/TravelTypeDto"
          },
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicUserDto"
              }
            ]
          }
        },
        "required": [
          "travelType",
          "user"
        ]
      },
...
      "MatchResultDto": {
        "type": "object",
        "properties": {
          "compatibilityScore": {
            "type": "object",
            "example": 78,
            "nullable": true
          },
          "available": {
            "type": "boolean",
            "example": true
          },
          "matchingAreas": {
            "example": [
              "카페",
              "자연"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "adjustmentAreas": {
            "example": [
              "쇼핑"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avoidAreas": {
            "example": [
              "클럽"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary": {
            "type": "string",
            "example": "자연·카페에서 잘 맞아요"
          }
        },
        "required": [
          "compatibilityScore",
          "available",
          "matchingAreas",
          "adjustmentAreas",
          "avoidAreas",
          "summary"
        ]
      },
...
      "AnalysisReportDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "roomId": {
            "type": "string"
          },
          "scheduleVersion": {
            "type": "number",
            "example": 7
          },
          "factsVersion": {
            "type": "number",
            "example": 12
          },
          "analysisVersion": {
            "type": "string",
            "example": "evidence-v2"
          },
          "routeAnalysis": {
            "type": "object",
            "example": {
              "available": true,
              "totalDistance": 12.4,
              "totalDurationSeconds": 1860,
              "knownSegmentCount": 1,
              "unknownSegmentCount": 0,
              "segments": [
                {
                  "from": "성산",
                  "to": "우도",
                  "status": "ok",
                  "distanceMeters": 8200,
                  "durationSeconds": 900,
                  "km": 8.2,
                  "source": "kakao-mobility"
                }
              ],
              "warnings": []
            }
          },
          "budgetAnalysis": {
            "type": "object",
            "example": {
              "available": false,
              "estimated": null,
              "reason": "cost_data_not_collected",
              "breakdown": []
            }
          },
          "densityAnalysis": {
            "type": "object",
            "example": {
              "method": "union_occupied_minutes_over_calendar_day",
              "scoreBasisMinutes": 1440,
              "byDay": [
                {
                  "day": 1,
                  "score": 16.7,
                  "occupiedMinutes": 240,
                  "freeMinutesWithinSpan": 30,
                  "spanMinutes": 270,
                  "message": "총 240분 일정, 일정 사이 여유 30분입니다."
                }
              ]
            }
          },
          "preferenceReflection": {
            "type": "object",
            "example": {
              "available": true,
              "score": 80,
              "method": "mean_member_preferred_tag_coverage",
              "confidence": "complete",
              "details": []
            }
          },
          "conflictAnalysis": {
            "type": "object",
            "example": {
              "overlaps": [],
              "closedVenues": [],
              "operatingHoursAvailable": false,
              "operatingHoursReason": "verified_operating_hours_not_available"
            }
          },
          "suggestions": {
            "example": [
              {
                "type": "route",
                "message": "동선을 줄여보세요"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stale": {
            "type": "boolean",
            "description": "현재 입력보다 오래된 리포트인지 여부"
          },
          "evidenceExpired": {
            "type": "boolean",
            "description": "교통 근거의 재조회 시간이 지난 리포트인지 여부"
          },
          "currentScheduleVersion": {
            "type": "number"
          },
          "currentFactsVersion": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "roomId",
          "scheduleVersion",
          "factsVersion",
          "analysisVersion",
          "routeAnalysis",
          "budgetAnalysis",
          "densityAnalysis",
          "preferenceReflection",
          "conflictAnalysis",
          "suggestions"
        ]
      },
...
    }
  }
}